requirements-dev.txt
boto3==1.26.20
debugpy==1.8.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-html==3.1.1
python-dotenv==0.17.0
requests==2.28.1
